docs: Add README.md
README for Final Project in MGMT646. Include a small how-to and
instructions to get started.

feat: Add file-compare.py script

This is the tool that I wrote as part of the Final Project
requirements. It will compare two files in various ways in order to
determine if the two files are the same.